feat(User): :sparkles: Add user registration and sidebar functionality.
Added a user registration page that allows users to create an account by entering their first name, last name, email, password, and confirming the password.  Also implemented a sidebar for authenticated users that provides navigation options such as viewing their history and logging out.  For users who are not logged in, the sidebar includes a link to the login page.

No breaking changes.